spi: Fix memory corruption bug in spi-litespi.c
Incrementing the tx_buf and rx_buf pointers in the SPI transfer
function can lead to memory corruption in the user of the SPI
device if it was expecting the pointeres to remain constant. This
was responsible for causing a kernel panic in the mmc-spi driver.

Gabriel Solmo, in his litex-rocket-rebase branch has already fixed
this bug, so I have really just merged his driver into the vexriscv
branch for the time being until I can get his kernel booting again
on a vexriscv platform.

Signed-off-by: Owen Kirby <oskirby@gmail.com>